+++ src/main/java/com/qdc/lims/dto/PurchaseRequest.java
@@ -1,5 +1,6 @@
 package com.qdc.lims.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -15,7 +16,7 @@ public record PurchaseRequest(
         Long supplierId,
         String invoiceNumber,
         List<PurchaseItemDTO> items,
-        Double amountPaidNow, // e.g. 5000.0
+        BigDecimal amountPaidNow, // e.g. 5000.0
         String paymentMode // "Cash", "Cheque", "Online"
 ) {
 }
